Implement prioritizeProvider ekam rule directive
This lets a script specify a tool as being suitable for host compilation
so that it's built before any codegen. Similarly, the requisite sources
are known to need compiling before the tool is attempted to be linked.

The pending actions list is split into 2 sets of prioritized queues.
The first keeps track of enqueued while the latter keeps track of
requeued. Requeued items are only processed again once the enqueued
items are complete. This ensures there's no infinite loops possible.